try to reduce code churn
